+++ trview.common/Strings.cpp
@@ -7,7 +7,7 @@ namespace trview
         std::vector<char> output(WideCharToMultiByte(CP_UTF8, 0, value.c_str(), -1, nullptr, 0, nullptr, nullptr), 0);
         if (!output.empty())
         {
-            WideCharToMultiByte(CP_UTF8, 0, value.c_str(), -1, &output[0], output.size(), nullptr, nullptr);
+            WideCharToMultiByte(CP_UTF8, 0, value.c_str(), -1, &output[0], static_cast<int>(output.size()), nullptr, nullptr);
         }
         return &output[0];
     }
@@ -17,7 +17,7 @@ namespace trview
         std::vector<wchar_t> output(MultiByteToWideChar(CP_UTF8, 0, value.c_str(), -1, nullptr, 0), 0);
         if (!output.empty())
         {
-            MultiByteToWideChar(CP_UTF8, 0, value.c_str(), -1, &output[0], output.size());
+            MultiByteToWideChar(CP_UTF8, 0, value.c_str(), -1, &output[0], static_cast<int>(output.size()));
         }
         return &output[0];
     }
